Refactored project name and folders.
