+++ __init__.py
@@ -1,0 +1,2 @@
+"""Initialize the package."""
+from ._local_regression.loess import Loess as Loess
