+++ _local_regression/__init__.py
@@ -1,0 +1,1 @@
+"""Initialize the _local_regression package."""
