+++ _local_regression/_not_fitted_error.py
