--- local_regression/__init__.py
@@ -1,1 +1,0 @@
-"""Initialize the local_regression package."""
